requirements.txt
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-collectionfield==0.0.4
django-phonenumber-field==2.0.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.0.4.1
django-toolbelt==0.0.1
djangorestframework==3.8.2
gunicorn==19.8.1
heroku==0.1.4
idna==2.6
phonenumberslite==8.9.5
Pillow==5.1.0
protobuf==3.5.2.post1
psycopg2==2.7.4
Pygments==2.2.0
pyodbc==4.0.23
python-dateutil==1.5
pytz==2018.4
requests==2.18.4
six==1.11.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
